Extracted utils for performing requests to separated class. Cleaned up
