+++ VoucherifyClient.php
@@ -1,116 +1,13 @@
 <?php
 
 namespace Voucherify {
-    
-    class VoucherifyClient {
-        
-        /**
-         * @var string
-         */
-        private static $apiURL = "https://api.voucherify.io/v1";
-        
-        /**
-         * @var string
-         */
-        private $apiID;
-        
-        /**
-         * @var string
-         */
-        private $apiKey;
-        
-        /**
-         * @param string $apiID
-         * @param string $apiKey
-         */
+
+    class VoucherifyClient extends VoucherifyRequest {
+
         public function __construct($apiID, $apiKey) {
-            
-            $this->setApiID($apiID);
-            $this->setApiKey($apiKey);
+            parent::__construct($apiID, $apiKey);
         }
-        
-        private function getHeaders() {
-            return [
-                "Content-Type: application/json",
-                "X-App-Id: " . $this->apiID,
-                "X-App-Token: " . $this->apiKey,
-                "X-Voucherify-Channel: PHP-SDK"
-            ];
-        }
-        
-        private function encodeParams($params) {
-            if (!is_array($params) && !is_object($params)) {
-                return $params;
-            }
-                
-            $result = array();
-            foreach ($params as $key => $value) {
-                if (is_null($value)) {
-                    continue;
-                }
-                $result[] = urlencode($key) . "=" . urlencode($value);
-            }
-            return implode("&", $result);
-        }
-        
-        /**
-         * @param string $method
-         * @param string $endpoint
-         * @param array|null $params
-         * @param string|array|object|null $data
-         *
-         * @throws Voucherify\ClientException
-         */
-        private function apiRequest($method, $endpoint, $params, $data) {
-            
-            $setParams  = $params   && in_array($method, [ "GET", "POST" ]);
-            $setData    = $data     && in_array($method, [ "POST", "PUT", "DELETE"]);
-            
-            $method = strtoupper($method);
-            $url = self::$apiURL . $endpoint . ($setParams ? "?" . $this->encodeParams($params) : "");
-           
-            $options = array();
-            $options[CURLOPT_URL]               = $url;
-            $options[CURLOPT_HTTPHEADER]        = $this->getHeaders();
-            $options[CURLOPT_RETURNTRANSFER]    = true; 
-            $options[CURLOPT_CUSTOMREQUEST]     = $method;
-            $options[CURLOPT_POSTFIELDS]        = $setData ? json_encode($data) : NULL;
-            
-            $curl = curl_init();
-            
-            curl_setopt_array($curl, $options);
-            
-            $result         = curl_exec($curl);
-            $statusCode     = curl_getinfo($curl, CURLINFO_HTTP_CODE);
-            $error          = curl_error($curl);
-            curl_close($curl);
-                
-            // Connection errors
-            if ($result === false) {
-                throw new ClientException($error);
-            }
-            // Invalid status code
-            else if ($statusCode >= 400) {
-                throw new ClientException("Unexpected status code: " . $statusCode . " - Details: " . $result);
-            }
-            
-            return json_decode($result);
-        }
-        
-        /**
-         * @param string $apiKey
-         */
-        public function setApiKey($apiKey) {
-            $this->apiKey = $apiKey;
-        }
-        
-        /**
-         * @param string $apiID
-         */
-        public function setApiID($apiID) {
-            $this->apiID = $apiID;
-        }
-        
+
         /**
          * @param string $code
          *
@@ -121,7 +18,7 @@ namespace Voucherify {
         public function get($code) {
             return $this->apiRequest("GET", "/vouchers/" . urlencode($code), NULL, NULL);
         }
-        
+
         /**
          * @param stdClass $voucher
          *
@@ -131,12 +28,12 @@ namespace Voucherify {
          */
         public function create($voucher) {
             if (isset($voucher->code)) {
-                return $this->apiRequest("POST", "/vouchers/" . urlencode($voucher->code), NULL, $voucher);    
+                return $this->apiRequest("POST", "/vouchers/" . urlencode($voucher->code), NULL, $voucher);
             } else {
-                return $this->apiRequest("POST", "/vouchers/", NULL, $voucher);    
+                return $this->apiRequest("POST", "/vouchers/", NULL, $voucher);
             }
         }
-        
+
         /**
          * @param string $code
          *
@@ -147,7 +44,7 @@ namespace Voucherify {
         public function enable($code) {
             return $this->apiRequest("POST", "/vouchers/" . urlencode($code) . "/enable", NULL, NULL);
         }
-        
+
         /**
          * @param string $code
          *
@@ -158,9 +55,9 @@ namespace Voucherify {
         public function disable($code) {
             return $this->apiRequest("POST", "/vouchers/" . urlencode($code) . "/disable", NULL, NULL);
         }
-        
+
         /**
-         * @param string $code 
+         * @param string $code
          *
          * Get voucher redemption
          *
@@ -169,10 +66,10 @@ namespace Voucherify {
         public function redemption($code) {
             return $this->apiRequest("GET", "/vouchers/" . urlencode($code) . "/redemption/", NULL, NULL);
         }
-        
+
         /**
-         * @param string|array $code Voucher code or array with voucher and customer items 
-         * @param string|null $trackingId Provided tracking id 
+         * @param string|array $code Voucher code or array with voucher and customer items
+         * @param string|null $trackingId Provided tracking id
          *
          * Redeem voucher
          *
@@ -184,12 +81,12 @@ namespace Voucherify {
                 $context = $code;
                 $code = $context["voucher"];
                 unset($context["voucher"]);
-            }        
-            return $this->apiRequest("POST", "/vouchers/" . urlencode($code) . "/redemption/", [ "tracking_id" => $trackingId ], $context);
+            }
+            return $this->apiRequest("POST", "/vouchers/" . urlencode($code) . "/redemption/", ["tracking_id" => $trackingId], $context);
         }
-        
+
         /**
-         * @param string $redemptionId 
+         * @param string $redemptionId
          * @param string|null $trackingId
          * @param string|null $reason
          *
@@ -199,9 +96,9 @@ namespace Voucherify {
          * @throws Voucherify\ClientException
          */
         public function rollback($redemptionId, $trackingId = NULL, $reason = NULL) {
-            return $this->apiRequest("POST", "/redemptions/" . urlencode($redemptionId) . "/rollback/", [ "tracking_id" => $trackingId, "reason" => $reason ], NULL);
+            return $this->apiRequest("POST", "/redemptions/" . urlencode($redemptionId) . "/rollback/", ["tracking_id" => $trackingId, "reason" => $reason], NULL);
         }
-        
+
         /**
          * @param array|stdClass $filter
          *
@@ -212,13 +109,13 @@ namespace Voucherify {
          * - campaign   - string
          * - category   - string
          * - customer   - string
-         * 
+         *
          * @throws Voucherify\ClientException
          */
         public function vouchers($filter) {
-            return $this->apiRequest("GET", "/vouchers/", $filter, NULL);    
+            return $this->apiRequest("GET", "/vouchers/", $filter, NULL);
         }
-        
+
         /**
          * @param array|stdClass $filter
          *
@@ -229,12 +126,11 @@ namespace Voucherify {
          * - end_date   - string (ISO8601 format, default is the end of current month)
          * - result     - string (Success|Failure-NotExist|Failure-Inactive)
          * - customer   - string
-         * 
+         *
          * @throws Voucherify\ClientException
          */
         public function redemptions($filter) {
-            return $this->apiRequest("GET", "/redemptions/", $filter, NULL);    
+            return $this->apiRequest("GET", "/redemptions/", $filter, NULL);
         }
     }
 }
-?>
